feat: compact FileOrder struct and eliminate per-pair mappings
State optimization reducing ~22-26 GB at 100M orders:

Step 1 - Compact FileOrder (6→3 slots):
- Remove FileMeta struct, flatten fileRoot into FileOrder
- Move uri to separate orderUri mapping (written once, read rarely)
- Remove price field (derived: escrow / numChunks / periods / replicas)
- Add _orderPrice() helper and getOrderPrice() public view
- Change startPeriod from uint64 to uint32
- Update placeOrder signature to flat args

Step 2 - Eliminate per-(order,node) pair mappings:
- Replace orderToNodes with orderAssignments (NodeAssignment struct)
- Pack node address + startPeriod in one slot per assignment
- Remove nodeOrderStartTimestamp and nodeOrderEarnings mappings
- Use nodeLastClaimPeriod watermark for reward settlement
- Compute earnings formulaically instead of storing running totals

292/292 tests pass. Contract sizes: FileMarket 21.9KB, Extension 24.0KB.

+++ test/market/MarketViews.t.sol
@@ -100,7 +100,7 @@ contract MarketViewsTest is MarketTestBase {
 
         _executeOrder(node1, orderId);
 
-        (, uint256 withdrawn, uint64 startPeriod, bool expired, address[] memory nodes) =
+        (, uint256 withdrawn, uint32 startPeriod, bool expired, address[] memory nodes) =
             marketExt.getOrderFinancials(orderId);
 
         assertEq(withdrawn, 0);
